+++ src/mkvtoolnix-gui/util/settings_names.h
@@ -74,6 +74,7 @@ char const * const s_valMergeAlwaysAddDroppedFiles                = "mergeAlways
 char const * const s_valMergeAlwaysCreateNewSettingsForVideoFiles = "mergeAlwaysCreateNewSettingsForVideoFiles";
 char const * const s_valMergeAlwaysShowOutputFileControls         = "mergeAlwaysShowOutputFileControls";
 char const * const s_valMergeAttachmentsAlwaysSkipForExistingName = "mergeAttachmentsAlwaysSkipForExistingName";
+char const * const s_valMergeDragAndDropDirectoriesPolicy         = "mergeDragAndDropDirectoriesPolicy";
 char const * const s_valMergeDragAndDropFilesPolicy               = "mergeDragAndDropFilesPolicy";
 char const * const s_valMergeEnableDialogNormGainRemoval          = "mergeEnableDialogNormGainRemoval";
 char const * const s_valMergeFileColors                           = "mergeFileColors";