+++ NEWS.md
@@ -9,6 +9,10 @@
   only want to relocate the preferences INI file, you can store the desired
   file name for the INI file in `MKVTOOLNIX_GUI_CONFIG_FILE`. Implements
   #3382.
+* MKVToolNix GUI: when adding directories via drag & drop/copy & paste the GUI
+  will now offer the user the option to create one multiplex settings tab per
+  directory & adding all files from that directory to the corresponding, newly
+  created tab. Implements #3371.
 
 ## Bug fixes